Fix/unit test case (#144)
* test: Update 'roo' to 'shenma' in snap file to prevent test cases from failing

This update changes the identifier 'roo' to 'shenma' in the snap file and code test file , ensuring that the relevant test cases now pass as expected.

* test: Fix failing test cases and update files under "e2e" to ensure the GitHub Action "integration-test" runs successfully

This commit includes fixes for existing failing test cases and necessary updates to files within the "e2e" directory to ensure that the "integration-test" workflow in GitHub Actions executes smoothly.

---------

Co-authored-by: dengbin <denbinbox@gmail.com>